adding modal to process list view
