+++ generics/components/organisms/Organism_ListAll.mjs
@@ -4,7 +4,6 @@ import {Atom_Input} from "../atoms/Atom_Input.mjs";
 import {Atom_ButtonPositive} from "../atoms/Atom_ButtonPositive.mjs";
 import {Molecule_List} from "../molecules/Molecule_List.mjs";
 import {Atom_Heading2} from "../atoms/Atom_Heading2.mjs";
-import { Modal_ProcessListAll } from "./Modal_ProcessListAll.mjs";
 
 export function Organism_ListAll (model){
 
@@ -22,11 +21,10 @@ export function Organism_ListAll (model){
                     </div>
                     
                 </div>
-                <div class="organism_list-all-search__lists">
+                <div id="organism-proc-lists" class="organism_list-all-search__lists">
                     ${slot('list1')}
                     ${slot('list2')}
                     ${slot('list3')}
-                    <div id="modal-processView"></div>
                 </div>
                 
             </div>
@@ -52,21 +50,6 @@ export function Organism_ListAll (model){
 
         let listComponent3 = new Molecule_List(model.lists3)
         this.fillSlot('list3', listComponent3.getElement())
-
-        this.getElement().querySelector(".organism_list-all-search__lists").addEventListener("click", (e) => {
-            console.log('btn-project button pressed')     
-
-            const modalId = document.getElementById('modal-processView')
-            
-            modalId.innerHTML = `
-                <div>
-                    ${slot("new-modal")}
-                </div>
-                `
-            this.modal = new Modal_ProcessListAll(model.content)
-
-            this.fillSlot("new-modal", this.modal.getElement());
-        });
 
     }
 }
