+++ generics/components/templates/Template_ListAllProcesses_Model.mjs
@@ -125,76 +125,76 @@ export function Template_ListAllProcesses_Model(model) {
                         },
                         items : model.items
                     },
-                    content : {
-                        organism_processModal : {
-                            molecule_modalProcParentBoxes: {
-                                atom_heading4a : {
-                                    text : model.boxParent1
-                                },
-                                atom_icon1 : {
-                                    icon : model.icon6
-                                },
-                                atom_heading4b : {
-                                    text : model.boxParent2
-                                },
-                                atom_icon2 : {
-                                    icon : model.icon7
-                                },
-                                atom_heading4c : {
-                                    text : model.boxParent3
-                                },
-                                atom_icon3 : {
-                                    icon : model.icon8
-                                }
-                            },
-                            molecule_modalInputProcOutput : {
-                                lists1 : {
-                                    atom_heading4 : {
-                                        text : model.listHeading4
-                                    },
-                                    items : model.items2
-                                },
-                                atom_icon1 : {
-                                    icon : model.icon9
-                                },
-                                atom_heading4 : {
-                                    text : model.centerText
-                                },
-                                atom_text1 : {
-                                    text : model.centerAbout
-                                },
-                                atom_icon2 : {
-                                    icon : model.icon10
-                                },
-                                lists2 : {
-                                    atom_heading4 : {
-                                        text : model.listHeading5
-                                    },
-                                    items : model.items3
-                                }
-                            },
-                            molecule_modalChildrenBoxes : {
-                                atom_icon1 : {
-                                    icon : model.icon11
-                                },
-                                atom_heading4a : {
-                                    text : model.boxChild1
-                                },
-                                atom_icon2 : {
-                                    icon : model.icon12
-                                },
-                                atom_heading4b : {
-                                    text : model.boxChild2
-                                },
-                                atom_icon3 : {
-                                    icon : model.icon13
-                                },
-                                atom_heading4c : {
-                                    text : model.boxChild3
-                                },
-                            }
-                        }
-                    }    
+                    // content : {
+                    //     organism_processModal : {
+                    //         molecule_modalProcParentBoxes: {
+                    //             atom_heading4a : {
+                    //                 text : model.boxParent1
+                    //             },
+                    //             atom_icon1 : {
+                    //                 icon : model.icon6
+                    //             },
+                    //             atom_heading4b : {
+                    //                 text : model.boxParent2
+                    //             },
+                    //             atom_icon2 : {
+                    //                 icon : model.icon7
+                    //             },
+                    //             atom_heading4c : {
+                    //                 text : model.boxParent3
+                    //             },
+                    //             atom_icon3 : {
+                    //                 icon : model.icon8
+                    //             }
+                    //         },
+                    //         molecule_modalInputProcOutput : {
+                    //             lists1 : {
+                    //                 atom_heading4 : {
+                    //                     text : model.listHeading4
+                    //                 },
+                    //                 items : model.items2
+                    //             },
+                    //             atom_icon1 : {
+                    //                 icon : model.icon9
+                    //             },
+                    //             atom_heading4 : {
+                    //                 text : model.centerText
+                    //             },
+                    //             atom_text1 : {
+                    //                 text : model.centerAbout
+                    //             },
+                    //             atom_icon2 : {
+                    //                 icon : model.icon10
+                    //             },
+                    //             lists2 : {
+                    //                 atom_heading4 : {
+                    //                     text : model.listHeading5
+                    //                 },
+                    //                 items : model.items3
+                    //             }
+                    //         },
+                    //         molecule_modalChildrenBoxes : {
+                    //             atom_icon1 : {
+                    //                 icon : model.icon11
+                    //             },
+                    //             atom_heading4a : {
+                    //                 text : model.boxChild1
+                    //             },
+                    //             atom_icon2 : {
+                    //                 icon : model.icon12
+                    //             },
+                    //             atom_heading4b : {
+                    //                 text : model.boxChild2
+                    //             },
+                    //             atom_icon3 : {
+                    //                 icon : model.icon13
+                    //             },
+                    //             atom_heading4c : {
+                    //                 text : model.boxChild3
+                    //             },
+                    //         }
+                    //     }
+                    // }    
                 },
                 organism_processModal : {
                 molecule_modalProcParentBoxes: {
@@ -263,7 +263,77 @@ export function Template_ListAllProcesses_Model(model) {
                         text : model.boxChild3
                     },
                 }
-            }
+                },
+                content : {
+                    organism_processModal : {
+                        molecule_modalProcParentBoxes: {
+                            atom_heading4a : {
+                                text : model.boxParent1
+                            },
+                            atom_icon1 : {
+                                icon : model.icon6
+                            },
+                            atom_heading4b : {
+                                text : model.boxParent2
+                            },
+                            atom_icon2 : {
+                                icon : model.icon7
+                            },
+                            atom_heading4c : {
+                                text : model.boxParent3
+                            },
+                            atom_icon3 : {
+                                icon : model.icon8
+                            }
+                        },
+                        molecule_modalInputProcOutput : {
+                            lists1 : {
+                                atom_heading4 : {
+                                    text : model.listHeading4
+                                },
+                                items : model.items2
+                            },
+                            atom_icon1 : {
+                                icon : model.icon9
+                            },
+                            atom_heading4 : {
+                                text : model.centerText
+                            },
+                            atom_text1 : {
+                                text : model.centerAbout
+                            },
+                            atom_icon2 : {
+                                icon : model.icon10
+                            },
+                            lists2 : {
+                                atom_heading4 : {
+                                    text : model.listHeading5
+                                },
+                                items : model.items3
+                            }
+                        },
+                        molecule_modalChildrenBoxes : {
+                            atom_icon1 : {
+                                icon : model.icon11
+                            },
+                            atom_heading4a : {
+                                text : model.boxChild1
+                            },
+                            atom_icon2 : {
+                                icon : model.icon12
+                            },
+                            atom_heading4b : {
+                                text : model.boxChild2
+                            },
+                            atom_icon3 : {
+                                icon : model.icon13
+                            },
+                            atom_heading4c : {
+                                text : model.boxChild3
+                            },
+                        }
+                    }
+                }   
             }
         }
