+++ generics/components/templates/Template_ListAllProcesses_View.mjs
@@ -4,6 +4,7 @@ import { slot } from "nd_frontend/core/helpers.mjs";
 import { Organism_Navbar } from "../organisms/Organism_Navbar.mjs";
 import { Organism_ListAll } from "../organisms/Organism_ListAll.mjs";
 import { Organism_ProcessModal } from "../organisms/Organism_ProcessModal.mjs";
+import { Modal_ProcessListAll } from "../organisms/Modal_ProcessListAll.mjs";
 
 export function Template_ListAllProcesses_View(view){
     
@@ -13,8 +14,8 @@ export function Template_ListAllProcesses_View(view){
         return `<div>
             ${slot("organismNavbar")}
             ${slot("organismListAll")}
-
             ${slot("modalprocess")}
+            <div id="modal-processView"></div>
         </div>`
     }
 
@@ -27,5 +28,20 @@ export function Template_ListAllProcesses_View(view){
         this.fillSlot("organismNavbar", organismNavbar.getElement())
         this.fillSlot("organismListAll", organism_listAll.getElement())
         this.fillSlot("modalprocess", modalprocess.getElement())
+
+        this.getElement().querySelector("#organism-proc-lists").addEventListener("click", (e) => {
+            console.log('btn-project button pressed')     
+
+            const modalId = document.getElementById('modal-processView')
+            
+            modalId.innerHTML = `
+                <div>
+                    ${slot("new-modal")}
+                </div>
+                `
+            this.modal = new Modal_ProcessListAll(model.content)
+
+            this.fillSlot("new-modal", this.modal.getElement());
+        });
     }
 }
